+++ spec/support/active_record.rb
@@ -20,6 +20,12 @@ ActiveRecord::Schema.define do
     t.string :bio
   end
 
+  # Second model, used by the multi-pagination specs
+  create_table :posts, force: true do |t|
+    t.string :title
+    t.timestamps
+  end
+
   # UUID-keyed table for primary-key tests
   create_table :uuid_records, primary_key: :uuid, id: :string, force: true do |t|
     t.string :label
@@ -34,6 +40,9 @@ class Profile < ActiveRecord::Base
   belongs_to :user
 end
 
+class Post < ActiveRecord::Base
+end
+
 class UuidRecord < ActiveRecord::Base
   self.primary_key = :uuid
 end
@@ -43,6 +52,7 @@ module ActiveRecordHelpers
     User.delete_all
     Profile.delete_all
     UuidRecord.delete_all
+    Post.delete_all
   end
 end